fix(security): Extract shared escapeHtml, fix XSS in aria-label (GOAP V6)
- Create public/js/utils/html.js shared escapeHtml module
- Remove duplicated escapeHtml from deal-card.js, deal-detail.js, deals.js
- Fix unescaped aria-label interpolation in deal-card.js
- Eliminates DRY violation across 3 frontend files
