+++ plans/GOAP-SWARM-V6-2026-07-07.md
@@ -1,0 +1,75 @@
+# GOAP Swarm Execution Plan V6 — 2026-07-07
+
+**Generated**: 2026-07-07
+**Strategy**: Parallel (3 independent file updates after shared module creation)
+**Source**: Verify command warnings + code quality audit
+**Skills Used**: goap-agent, typescript-coding-standards
+**Status**: ✅ COMPLETED
+
+---
+
+## Task Analysis
+
+**Primary Goal**: Fix all verify warnings and pre-existing code quality issues
+**Constraints**: No breaking changes; maintain existing functionality
+**Complexity**: Low-Medium (DRY extraction + XSS fix)
+
+### Issues Found
+
+| # | Issue | Severity | Files Affected |
+|:---|:---|:---|:---|
+| 1 | Duplicated `escapeHtml` in 3 files | Medium | deal-card.js, deal-detail.js, deals.js |
+| 2 | Unescaped `aria-label` interpolation | Medium | deal-card.js:80 |
+| 3 | Duplicated `HTML_ESCAPES`/`HTML_ENTITIES` constants | Low | deal-card.js, deal-detail.js, deals.js |
+
+---
+
+## Execution Plan
+
+**Strategy**: Sequential (create shared module) → Parallel (update 3 files)
+
+```
+┌─────────────────────────────────────────────────────────────────┐
+│                    GOAP Swarm Controller                        │
+│  Strategy: Sequential → Parallel                                │
+└─────────────────────────┬───────────────────────────────────────┘
+                          │
+              ┌───────────────────────┐
+              │  Create shared module │
+              │  public/js/utils/     │
+              │  html.js              │
+              └───────────────────────┘
+                          │
+          ┌───────────────┼───────────────┐
+          ▼               ▼               ▼
+    ┌──────────┐    ┌──────────┐    ┌──────────┐
+    │ Agent 1  │    │ Agent 2  │    │ Agent 3  │
+    │ deal-    │    │ deal-    │    │ deals.js │
+    │ card.js  │    │ detail.js│    │ Update   │
+    │ Update   │    │ Update   │    │          │
+    └──────────┘    └──────────┘    └──────────┘
+          │               │               │
+          └───────────────┼───────────────┘
+                          ▼
+              ┌───────────────────────┐
+              │  Verify (lint + fmt)  │
+              └───────────────────────┘
+                          │
+                          ▼
+              ┌───────────────────────┐
+              │  Commit + Push        │
+              └───────────────────────┘
+```
+
+---
+
+## Quality Gates
+
+1. No duplicate `escapeHtml` definitions
+2. All `aria-label` values escaped
+3. Prettier formatting passes
+4. No lint warnings
+
+---
+
+*GOAP Swarm V6 plan for fixing verify warnings and code quality issues.*
