+++ public/js/utils/html.js
@@ -1,0 +1,12 @@
+const HTML_ESCAPES = {
+  "&": "&amp;",
+  "<": "&lt;",
+  ">": "&gt;",
+  '"': "&quot;",
+  "'": "&#39;",
+};
+
+export function escapeHtml(value) {
+  if (value == null) return "";
+  return String(value).replace(/[&<>"']/g, (ch) => HTML_ESCAPES[ch] || ch);
+}
